clang-tidy: modernize-use-override
Adds override (introduced in C++11) to overridden virtual functions and
removes virtual from those functions as it is not required.

virtual on non base class implementations was used to help indicate to
the user that a function was virtual. C++ compilers did not use the
presence of this to signify an overridden function.

In C++11 override and final keywords were introduced to allow overridden
functions to be marked appropriately. Their presence allows compilers to
verify that an overridden function correctly overrides a base class
implementation.

This can be useful as compilers can generate a compile time error when:
- The base class implementation function signature changes.
- The user has not created the override with the correct signature.