Use fixed-width integer types where appropriate (#17)
This changes some of the functions to use `int32_t` instead of `int` or `long` in places where it is specifically operating on four bytes. It also replaces hardcoded sizes with more readable `sizeof` expressions (e.g. `sizeof(float)` instead of magic number `4`).

Since this changes the signatures of a few functions, this is a breaking change on any platform where `sizeof(long) != 8 || sizeof(int) != 4`. I have provided `store_long()` as a shim with deprecation for `store_int32` so at least the old function name is still available.

rbdannenberg's portsmf has a slightly different implementation that utilizes 64-bit integers, which may be worth implementing here, but for now I have only made changes to the existing functions.

This also *should* remove the need for some msvc warning pragmas by using `uintptr_t` instead of `long` when doing casting for pointer arithmetic, though I am unable to test this myself for the time being.

Reviewed-on: https://codeberg.org/tenacityteam/portsmf/pulls/17
Reviewed-by: Avery King <gperson@disroot.org>
Co-authored-by: Fawn Sannar <rubiefawn@gmail.com>
Co-committed-by: Fawn Sannar <rubiefawn@gmail.com>