Add x64/arm variants for Github actions
This does change the artifact names so that they use the runner name
instead of the job name. This is mainly to avoid duplicating jobs for
each platform. However, if this is undesirable, the artifact name can
easily be changed to use the job name again, and each runner can be
given its own job again.