clang-tidy: bugprone-branch-clone
Checks for repeated branches in if/else if/else chains, consecutive
repeated branches in switch statements and identical true and false
branches in conditional operators.

This commit fixes two such instances. One of these was a false positive
due to an empty branch. The comments say that branch should never happen
and that it might be useful to assert(false) rather than do nothing. I
added the assertion accordingly, which also got rid of the false
positive.